pyproject.toml
[tool.poetry]
name = "avocadosh.xyz"
version = "0.1.0"
description = ""
authors = ["Maor Kadosh <git@avocadosh.xyz>"]

[tool.poetry.dependencies]
python = "^3.8"
Jinja2 = "^3.0.2"
Pillow = "^8.4.0"
pylast = "^4.3.0"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
